feat: add `--identifier` flag and prompt
closes #750
