+++ .changes/identifier.md
@@ -1,0 +1,7 @@
+---
+"create-tauri-app": "minor"
+"create-tauri-app-js": "minor"
+---
+
+Add `--identifier` flag and prompt to choose the identifier of your app.
+
