+++ templates/_base_/src-tauri/%(stable)%tauri.conf.json.lte
@@ -31,7 +31,7 @@
     "bundle": {
       "active": true,
       "targets": "all",
-      "identifier": "com.tauri.dev",
+      "identifier": "{% identifier %}",
       "icon": [
         "icons/32x32.png",
         "icons/128x128.png",
